Create Bubble.py
Bubble Sort is one of the simplest sorting algorithms. It repeatedly steps through the list to be sorted, compares adjacent elements, and swaps them if they are in the wrong order. The pass through the list is repeated until no swaps are needed, which indicates that the list is sorted